Create DropDown component and update screens.
